deps: WASI-Virt with upstreamed walrus changes
After the work done upstreaming operations used in `src/walrus_ops.rs`
to walrus, the code for WASI-Virt can be updated to use the new APIs
and reduce custom code.

This commit updates the WASI-Virt codebase to use the upstreamed
functions and adds some light refactors to the more mechanical pieces
of code for stubbing/denying/stripping.

see: https://github.com/bytecodealliance/WASI-Virt/issues/20
see: https://github.com/rustwasm/walrus/pull/252
