--- src/virt_deny.rs
@@ -1,771 +1,0 @@
-use anyhow::Result;
-use walrus::{Module, ValType};
-
-use crate::{
-    virt_io::{stub_clocks_virt, stub_http_virt, stub_sockets_virt},
-    walrus_ops::add_stub_exported_func,
-};
-
-// set exports to deny clock access
-pub(crate) fn deny_clocks_virt(module: &mut Module) -> Result<()> {
-    stub_clocks_virt(module)?;
-    add_stub_exported_func(
-        module,
-        "wasi:clocks/monotonic-clock#now",
-        vec![],
-        vec![ValType::I64],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:clocks/monotonic-clock#resolution",
-        vec![],
-        vec![ValType::I64],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:clocks/monotonic-clock#subscribe",
-        vec![ValType::I64, ValType::I32],
-        vec![ValType::I32],
-    )?;
-
-    add_stub_exported_func(
-        module,
-        "wasi:clocks/wall-clock#now",
-        vec![],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:clocks/wall-clock#resolution",
-        vec![],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:clocks/wall-clock#subscribe",
-        vec![ValType::I64, ValType::I32],
-        vec![ValType::I32],
-    )?;
-
-    add_stub_exported_func(
-        module,
-        "wasi:clocks/timezone#display",
-        vec![ValType::I32, ValType::I64, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "cabi_post_wasi:clocks/timezone#display",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:clocks/timezone#utc-offset",
-        vec![ValType::I32, ValType::I64, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:clocks/timezone#drop-timezone",
-        vec![ValType::I32],
-        vec![],
-    )?;
-
-    Ok(())
-}
-
-pub(crate) fn deny_http_virt(module: &mut Module) -> Result<()> {
-    stub_http_virt(module)?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/incoming-handler#handle",
-        vec![ValType::I32, ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/outgoing-handler#handle",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#drop-fields",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#new-fields",
-        vec![ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#fields-get",
-        vec![ValType::I32, ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#fields-set",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#fields-delete",
-        vec![ValType::I32, ValType::I32, ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#fields-append",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#fields-entries",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#fields-clone",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#finish-incoming-stream",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#finish-outgoing-stream",
-        vec![ValType::I32, ValType::I32, ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#drop-incoming-request",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#drop-outgoing-request",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#incoming-request-method",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#incoming-request-path-with-query",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#incoming-request-scheme",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#incoming-request-authority",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#incoming-request-headers",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#incoming-request-consume",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#new-outgoing-request",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#outgoing-request-write",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#drop-response-outparam",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#set-response-outparam",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#drop-incoming-response",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#drop-outgoing-response",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#incoming-response-status",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#incoming-response-headers",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#incoming-response-consume",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#new-outgoing-response",
-        vec![ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#outgoing-response-write",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#drop-future-incoming-response",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#future-incoming-response-get",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:http/types#listen-to-future-incoming-response",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    Ok(())
-}
-
-pub(crate) fn deny_random_virt(module: &mut Module) -> Result<()> {
-    add_stub_exported_func(
-        module,
-        "wasi:random/random#get-random-bytes",
-        vec![ValType::I64],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "cabi_post_wasi:random/random#get-random-bytes",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:random/random#get-random-u64",
-        vec![],
-        vec![ValType::I64],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:random/insecure#get-insecure-random-bytes",
-        vec![ValType::I64],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "cabi_post_wasi:random/insecure#get-insecure-random-bytes",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:random/insecure#get-insecure-random-u64",
-        vec![],
-        vec![ValType::I64],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:random/insecure-seed#insecure-seed",
-        vec![],
-        vec![ValType::I32],
-    )?;
-    Ok(())
-}
-
-pub(crate) fn deny_exit_virt(module: &mut Module) -> Result<()> {
-    add_stub_exported_func(module, "wasi:cli/exit#exit", vec![ValType::I32], vec![])?;
-    Ok(())
-}
-
-pub(crate) fn deny_sockets_virt(module: &mut Module) -> Result<()> {
-    stub_sockets_virt(module)?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/network#drop-network",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/instance-network#instance-network",
-        vec![],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/ip-name-lookup#resolve-addresses",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/ip-name-lookup#resolve-next-address",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/ip-name-lookup#drop-resolve-address-stream",
-        vec![ValType::I32],
-        vec![],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/ip-name-lookup#subscribe",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp-create-socket#create-tcp-socket",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#start-bind",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#finish-bind",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#start-connect",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#finish-connect",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#start-listen",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#finish-listen",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#accept",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#local-address",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#remote-address",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#address-family",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#ipv6-only",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#set-ipv6-only",
-        vec![ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#set-listen-backlog-size",
-        vec![ValType::I32, ValType::I64],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#keep-alive",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#set-keep-alive",
-        vec![ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#no-delay",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#set-no-delay",
-        vec![ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#unicast-hop-limit",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#set-unicast-hop-limit",
-        vec![ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#receive-buffer-size",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#set-receive-buffer-size",
-        vec![ValType::I32, ValType::I64],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#send-buffer-size",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#set-send-buffer-size",
-        vec![ValType::I32, ValType::I64],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#subscribe",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#shutdown",
-        vec![ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/tcp#drop-tcp-socket",
-        vec![ValType::I32],
-        vec![],
-    )?;
-
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp-create-socket#create-udp-socket",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#start-bind",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#finish-bind",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#start-connect",
-        vec![
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-            ValType::I32,
-        ],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#finish-connect",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#receive",
-        vec![ValType::I32, ValType::I64],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#send",
-        vec![ValType::I32, ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#local-address",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#remote-address",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#address-family",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#ipv6-only",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#set-ipv6-only",
-        vec![ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#unicast-hop-limit",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#set-unicast-hop-limit",
-        vec![ValType::I32, ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#receive-buffer-size",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#set-receive-buffer-size",
-        vec![ValType::I32, ValType::I64],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#send-buffer-size",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#set-send-buffer-size",
-        vec![ValType::I32, ValType::I64],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#subscribe",
-        vec![ValType::I32],
-        vec![ValType::I32],
-    )?;
-    add_stub_exported_func(
-        module,
-        "wasi:sockets/udp#drop-udp-socket",
-        vec![ValType::I32],
-        vec![],
-    )?;
-
-    Ok(())
-}
